+++ dhis-2/dhis-services/dhis-service-analytics/src/main/java/org/hisp/dhis/analytics/table/JdbcEnrollmentAnalyticsTableManager.java
@@ -27,6 +27,10 @@
  */
 package org.hisp.dhis.analytics.table;
 
+import static org.hisp.dhis.analytics.table.model.Skip.SKIP;
+import static org.hisp.dhis.analytics.util.AnalyticsUtils.getClosingParentheses;
+import static org.hisp.dhis.analytics.util.AnalyticsUtils.getColumnType;
+import static org.hisp.dhis.db.model.DataType.TEXT;
 import static org.hisp.dhis.util.DateUtils.toLongDate;
 
 import java.util.ArrayList;
@@ -37,14 +41,17 @@ import org.hisp.dhis.analytics.AnalyticsTableHookService;
 import org.hisp.dhis.analytics.AnalyticsTableType;
 import org.hisp.dhis.analytics.AnalyticsTableUpdateParams;
 import org.hisp.dhis.analytics.partition.PartitionManager;
+import org.hisp.dhis.analytics.table.model.AnalyticsDimensionType;
 import org.hisp.dhis.analytics.table.model.AnalyticsTable;
 import org.hisp.dhis.analytics.table.model.AnalyticsTableColumn;
 import org.hisp.dhis.analytics.table.model.AnalyticsTablePartition;
+import org.hisp.dhis.analytics.table.model.Skip;
 import org.hisp.dhis.analytics.table.setting.AnalyticsTableSettings;
 import org.hisp.dhis.category.CategoryService;
 import org.hisp.dhis.common.IdentifiableObjectManager;
 import org.hisp.dhis.commons.collection.UniqueArrayList;
 import org.hisp.dhis.dataapproval.DataApprovalLevelService;
+import org.hisp.dhis.db.model.DataType;
 import org.hisp.dhis.db.model.Logged;
 import org.hisp.dhis.db.sql.SqlBuilder;
 import org.hisp.dhis.organisationunit.OrganisationUnitService;
@@ -53,6 +60,7 @@ import org.hisp.dhis.program.Program;
 import org.hisp.dhis.resourcetable.ResourceTableService;
 import org.hisp.dhis.setting.SystemSettingsProvider;
 import org.hisp.dhis.system.database.DatabaseInfoProvider;
+import org.hisp.dhis.trackedentity.TrackedEntityAttribute;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
@@ -188,5 +196,61 @@ public class JdbcEnrollmentAnalyticsTableManager extends AbstractEventJdbcTableM
     }
 
     return filterDimensionColumns(columns);
+  }
+
+  private List<AnalyticsTableColumn> getTrackedEntityAttributeColumns(Program program) {
+    List<AnalyticsTableColumn> columns = new ArrayList<>();
+
+    for (TrackedEntityAttribute attribute : program.getNonConfidentialTrackedEntityAttributes()) {
+      DataType dataType = getColumnType(attribute.getValueType(), isSpatialSupport());
+      String dataClause =
+          attribute.isNumericType()
+              ? getNumericClause()
+              : attribute.isDateType() ? getDateClause() : "";
+      String select = getSelectExpressionForAttribute(attribute.getValueType(), "value");
+      Skip skipIndex = skipIndex(attribute.getValueType(), attribute.hasOptionSet());
+
+      String sql =
+          replaceQualify(
+              """
+              (select ${select} from ${trackedentityattributevalue} \
+              where trackedentityid=en.trackedentityid \
+              and trackedentityattributeid=${attributeId}\
+              ${dataClause})${closingParentheses} as ${attributeUid}""",
+              Map.of(
+                  "select",
+                  select,
+                  "attributeId",
+                  String.valueOf(attribute.getId()),
+                  "dataClause",
+                  dataClause,
+                  "closingParentheses",
+                  getClosingParentheses(select),
+                  "attributeUid",
+                  quote(attribute.getUid())));
+      columns.add(
+          AnalyticsTableColumn.builder()
+              .name(attribute.getUid())
+              .dimensionType(AnalyticsDimensionType.DYNAMIC)
+              .dataType(dataType)
+              .selectExpression(sql)
+              .skipIndex(skipIndex)
+              .build());
+
+      if (attribute.getValueType().isOrganisationUnit()) {
+        String fromTypeSql = "ou.name from organisationunit ou where ou.uid = (select value";
+        String ouNameSql = selectForInsert(attribute, fromTypeSql, dataClause);
+
+        columns.add(
+            AnalyticsTableColumn.builder()
+                .name((attribute.getUid() + OU_NAME_COL_SUFFIX))
+                .dimensionType(AnalyticsDimensionType.DYNAMIC)
+                .dataType(TEXT)
+                .selectExpression(ouNameSql)
+                .skipIndex(SKIP)
+                .build());
+      }
+    }
+    return columns;
   }
 }
